+++ test/nn/conv/test_message_passing.py
@@ -15,7 +15,12 @@ from torch_geometric.typing import (
     Size,
     SparseTensor,
 )
-from torch_geometric.utils import scatter, spmm, to_torch_csc_tensor
+from torch_geometric.utils import (
+    add_self_loops,
+    scatter,
+    spmm,
+    to_torch_csc_tensor,
+)
 
 
 class MyConv(MessagePassing):
@@ -51,6 +56,17 @@ class MyConv(MessagePassing):
     def message_and_aggregate(self, adj_t: SparseTensor,
                               x: OptPairTensor) -> Tensor:
         return spmm(adj_t, x[0], reduce=self.aggr)
+
+
+class MyConvWithSelfLoops(MessagePassing):
+    def __init__(self, aggr: str = 'add'):
+        super().__init__(aggr=aggr)
+
+    def forward(self, x: Tensor, edge_index: torch.Tensor) -> Tensor:
+        edge_index, _ = add_self_loops(edge_index)
+
+        # propagate_type: (x: Tensor)
+        return self.propagate(edge_index, x=x, size=None)
 
 
 def test_my_conv_basic():
@@ -608,49 +624,28 @@ def test_message_passing_int32_edge_index():
     assert conv(x, edge_index, edge_weight).size() == (4, 32)
 
 
-class MyConvWithSelfLoops(MessagePassing):
-    def __init__(self, aggr: str = 'add'):
-        super().__init__(aggr=aggr)
-
-    def forward(self, x: Tensor, edge_index: torch.Tensor) -> Tensor:
-        edge_index, _ = torch_geometric.utils.add_self_loops(edge_index)
-
-        # propagate_type: (x: Tensor)
-        return self.propagate(edge_index, x=x, size=None)
-
-    def message(self, x_j: Tensor) -> Tensor:
-        return x_j
-
-
-@pytest.mark.parametrize('forward_num_nodes', [4, 8, 2, 0])
-def test_my_conv_with_self_loops_traceable(forward_num_nodes):
-    """
-    torch.jit.trace a MessagePassing layer which adds self loops and test it
-    with different input sizes: more nodes, less nodes and no nodes.
-    All three cases are tested because all previously gave wrong results
-    or raised an error, for different reasons, all caused by a constant
-    number of nodes being traced in add_self_loops().
-    """
-    trace_num_nodes = 4
-    trace_x = torch.randn(trace_num_nodes, 16)
-    trace_edge_index = torch.stack([
-        torch.arange(0, trace_num_nodes - 1),
-        torch.arange(1, trace_num_nodes)
-    ])
-    forward_x = torch.randn(forward_num_nodes, 16)
-    forward_edge_index = (torch.stack([
-        torch.arange(0, forward_num_nodes - 1),
-        torch.arange(1, forward_num_nodes)
-    ]) if forward_num_nodes > 0 else torch.zeros((2, 0), dtype=torch.long))
+@pytest.mark.parametrize('num_nodes', [4, 8, 2, 0])
+def test_traceable_my_conv_with_self_loops(num_nodes):
+    # `torch.jit.trace` a `MessagePassing` layer that adds self loops and test
+    # it across different input sizes.
+    x = torch.randn(4, 16)
+    edge_index = torch.tensor([[0, 1, 1, 2, 2, 3], [1, 0, 2, 1, 3, 2]])
 
     conv = MyConvWithSelfLoops()
-
-    traced_conv = torch.jit.trace(conv, ((trace_x, trace_edge_index)))
+    traced_conv = torch.jit.trace(conv, ((x, edge_index)))
     scripted_conv = torch.jit.script(conv.jittable())
 
-    out = conv.forward(forward_x, forward_edge_index)
-    out_traced = traced_conv.forward(forward_x, forward_edge_index)
-    out_scripted = scripted_conv.forward(forward_x, forward_edge_index)
+    x = torch.randn(num_nodes, 16)
+    if num_nodes > 0:
+        edge_index = torch.stack(
+            [torch.arange(0, num_nodes - 1),
+             torch.arange(1, num_nodes)], dim=0)
+    else:
+        edge_index = torch.empty((2, 0), dtype=torch.long)
 
-    assert (torch.allclose(out, out_traced))
-    assert (torch.allclose(out, out_scripted))
+    out = conv(x, edge_index)
+    traced_out = traced_conv(x, edge_index)
+    scripted_out = scripted_conv(x, edge_index)
+
+    assert torch.allclose(out, traced_out)
+    assert torch.allclose(out, scripted_out)